+++ tests/conftest.py
@@ -4,4 +4,4 @@ import os
 sys.path.append(os.path.join(os.path.dirname(__file__), 'fixtures'))
 
 from .fixtures.script_scraper_fixtures import *
-from .fixtures.word_and_line_counts_fixtures import *
+from .fixtures.utils_fixtures import *